Move DataMixingSiStripMCDigiWorker to PremixingSiStripWorker
